fix: remove env tampering entirely
Since env vars take precedence on the default credentials chain, this
directly interferes with existing configurations, such as those on the
credentials file.

The failure mode of this change is that tests will fail if there is no
valid credential to be found, and the solution is then to set those on
the env beforehand.